Formatting edits
I edited both the register and login pages. For the register page,
I bolded the "Enter ___" lines. For the login page, I formatted
it the same as the register page (dark header, white text, icon
on the left). I also changed the width of the form entry lines
so that it didn't span the entire page width.